Limit drush cleanup to D8 platforms in static — fixes #1378
Otherwise it will delete some system drush files once /var/aegir/hostmaster-* platform is verified because on master instance drush is located on the same level in /var/aegir/drush

+++ aegir/tools/bin/fix-drupal-platform-ownership.sh
@@ -65,14 +65,19 @@ cd ${drupal_root}
 printf "Setting ownership of "${drupal_root}" to: user => "${script_user}" group => "users"\n"
 chown ${script_user}:users ${drupal_root}
 mkdir -p ${drupal_root}/sites/all/{modules,themes,libraries,drush}
-rm -f ${drupal_root}/../vendor/bin/drush*
-rm -f ${drupal_root}/vendor/bin/drush*
-rm -f ${drupal_root}/../drush/*
-rm -f ${drupal_root}/sites/development.services.yml
+if [[ "${drupal_root}" =~ "/static/" ]] && [ -e "${drupal_root}/core" ]; then
+  rm -f ${drupal_root}/../vendor/bin/drush*
+  rm -f ${drupal_root}/vendor/bin/drush*
+  rm -f ${drupal_root}/../drush/*
+  rm -f ${drupal_root}/sites/development.services.yml
+fi
 chown -R ${script_user}:users \
   ${drupal_root}/sites/all/{modules,themes,libraries,includes,misc,profiles,core,vendor,drush}/*
-chown -R ${script_user}:users ${drupal_root}/../vendor/*
-chown -R ${script_user}:users ${drupal_root}/../drush/*
+if [[ "${drupal_root}" =~ "/static/" ]] && [ -e "${drupal_root}/core" ]; then
+  chown -R ${script_user}:users ${drupal_root}/../vendor/*
+  chown -R ${script_user}:users ${drupal_root}/../drush/*
+  chown ${script_user}:users ${drupal_root}/../vendor
+fi
 chown ${script_user}:users \
   ${drupal_root}/sites/all/drush/drushrc.php \
   ${drupal_root}/sites \
@@ -80,8 +85,7 @@ chown ${script_user}:users \
   ${drupal_root}/sites/sites.php \
   ${drupal_root}/sites/all \
   ${drupal_root}/sites/all/{modules,themes,libraries,drush} \
-  ${drupal_root}/{modules,themes,libraries,includes,misc,profiles,core,vendor} \
-  ${drupal_root}/../vendor
+  ${drupal_root}/{modules,themes,libraries,includes,misc,profiles,core,vendor}
 
 ### known exceptions
 chown -R ${script_user}:www-data \
